+++ src/main/java/org/kumoricon/model/order/Payment.java
@@ -4,22 +4,26 @@ package org.kumoricon.model.order;
 import org.hibernate.annotations.NotFound;
 import org.hibernate.annotations.NotFoundAction;
 import org.kumoricon.model.Record;
+import org.kumoricon.model.session.Session;
 import org.kumoricon.model.user.User;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.ManyToOne;
+import javax.persistence.Table;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
+@Table(name="payments")
 public class Payment extends Record {
     @NotNull
     @Min(0)
     private BigDecimal amount;
 
+    @NotNull
     @ManyToOne
     private Order order;
 
@@ -33,24 +37,40 @@ public class Payment extends Record {
     private String paymentLocation;
     private String authNumber;
 
+    @NotNull
+    @ManyToOne
+    private Session session;
+
     public enum PaymentType {
-        CASH {
+        CASH(0) {
             public String toString() {
                 return "Cash";
             }
-        }, CHECK {
+        },
+        CHECK(1) {
             public String toString() {
                 return "Check/Money Order";
             }
-        }, CREDIT {
+        },
+        CREDIT(2) {
             public String toString() {
                 return "Credit Card";
             }
-        }, PREREG {
+        },
+        PREREG(3) {
             public String toString() {
                 return "Pre Reg";
             }
         };
+
+        private final int value;
+        private PaymentType(int value) {
+            this.value = value;
+        }
+
+        public int getValue() {
+            return value;
+        }
 
         public static PaymentType fromInteger(Integer typeId) {
             PaymentType[] paymentTypes = PaymentType.values();
@@ -102,6 +122,9 @@ public class Payment extends Record {
 
     public String getAuthNumber() { return authNumber; }
     public void setAuthNumber(String authNumber) { this.authNumber = authNumber; }
+
+    public Session getSession() { return session; }
+    public void setSession(Session session) { this.session = session; }
 
     public String toString() {
         if (id != null) {