+++ src/main/java/org/kumoricon/model/user/UserRepository.java
@@ -7,7 +7,6 @@ import java.util.List;
 
 @Service
 public interface UserRepository extends JpaRepository<User, Integer> {
-
     List<User> findByLastNameStartsWithIgnoreCase(String lastName);
     User findOneByUsernameIgnoreCase(String username);
 }