+++ src/main/java/org/kumoricon/service/validate/PaymentValidator.java
@@ -11,16 +11,16 @@ public class PaymentValidator extends Validator {
     public static boolean validate (Payment payment) {
         if (payment.getPaymentType() == Payment.PaymentType.CREDIT) {
             if (isNullOrEmpty(payment.getAuthNumber())) {
-                throw new ValueException("Credit cards must have authorization number");
+                throw new ValueException("Credit cards must have authorization number in note field");
             }
         } else if (payment.getPaymentType() == Payment.PaymentType.CHECK) {
             if (isNullOrEmpty(payment.getAuthNumber())) {
-               throw new ValueException("Checks must have a check number");
+               throw new ValueException("Checks must have a check number in note field");
             }
         }
 
         if (payment.getAmount() == null || BigDecimal.ZERO.compareTo(payment.getAmount()) > 0) {
-            throw new ValueException("Amount must not be negative");
+            throw new ValueException("Amount must not be negative or empty");
         }
 
         return true;