+++ src/main/java/org/kumoricon/site/LogoutPresenter.java
@@ -2,11 +2,15 @@ package org.kumoricon.site;
 
 import org.kumoricon.KumoRegUI;
 import org.kumoricon.model.order.OrderRepository;
+import org.kumoricon.model.session.Session;
+import org.kumoricon.model.session.SessionService;
 import org.kumoricon.model.user.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+
+import java.math.BigDecimal;
 
 @Controller
 public class LogoutPresenter {
@@ -14,6 +18,9 @@ public class LogoutPresenter {
 
     @Autowired
     OrderRepository orderRepository;
+
+    @Autowired
+    SessionService sessionService;
 
     public LogoutPresenter() {
     }
@@ -23,18 +30,23 @@ public class LogoutPresenter {
      * @param user User
      * @return True if session has orders worth more than $0.
      */
-    public Boolean tillReortNotPrinted(User user) {
+    public Boolean tillReportNotPrinted(User user) {
         if (user == null) {
-            log.error("tillReortNotPrinted() called with null User");
+            log.error("tillReportNotPrinted() called with null User");
             return false;
         }
 
-        Float total = orderRepository.getSessionOrderTotal(user.getId(), user.getSessionNumber());
-        if (total == null || total.compareTo(0.0f) <= 0) {
-            return false;
+        if (sessionService.userHasOpenSession(user)) {
+            Session currentSession = sessionService.getCurrentSessionForUser(user);
+            BigDecimal total = sessionService.getTotalForSession(currentSession);
+            if (total == null || total.compareTo(BigDecimal.ZERO) <= 0) {
+                return false;
+            } else {
+                log.info("{} is logging out but needs to print the till report", user);
+                return true;
+            }
         } else {
-            log.info("{} is logging out but needs to print the till report", user);
-            return true;
+            return false;
         }
     }