+++ src/main/java/org/kumoricon/site/utility/closeouttill/CloseOutTillPresenter.java
@@ -2,6 +2,8 @@ package org.kumoricon.site.utility.closeouttill;
 
 import org.kumoricon.model.order.OrderRepository;
 import org.kumoricon.model.order.Payment;
+import org.kumoricon.model.session.Session;
+import org.kumoricon.model.session.SessionService;
 import org.kumoricon.model.user.User;
 import org.kumoricon.model.user.UserRepository;
 import org.kumoricon.service.print.ReportPrintService;
@@ -12,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Controller;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -26,25 +29,26 @@ public class CloseOutTillPresenter {
     @Autowired
     private ReportPrintService reportPrintService;
 
+    @Autowired
+    private SessionService sessionService;
+
     private static final Logger log = LoggerFactory.getLogger(CloseOutTillPresenter.class);
 
     public void closeTill(CloseOutTillView view, User currentUser) {
         if (currentUser != null) {
-            Integer sessionNumber = currentUser.getSessionNumber();
-            log.info("{} closing out till, session number {}", currentUser, sessionNumber);
-            StringBuilder output = new StringBuilder();
-            List<Object[]> results = orderRepository.getSessionOrderCountsAndTotals(
-                    currentUser.getId(), sessionNumber);
-            String report = TillReportPresenter.getTillReportStr(currentUser, sessionNumber, results);
-            output.append(report);
-            log.info("Till report:\n" + report);
-            sessionNumber += 1;
-            currentUser.setSessionNumber(sessionNumber);
-            userRepository.save(currentUser);
-            output.append(String.format("Session closed. New session number is: %d", sessionNumber));
-            view.showData(output.toString());
-            log.info("{} created new till session, number {}", currentUser, currentUser.getSessionNumber());
-            view.notify(reportPrintService.printReport(output.toString(), view.getCurrentClientIPAddress()));
+            if (sessionService.userHasOpenSession(currentUser)) {
+                Session currentSession = sessionService.getCurrentSessionForUser(currentUser);
+                log.info("{} closing out till, session number {}", currentUser, currentSession);
+                currentSession = sessionService.closeSessionForUser(currentUser);
+
+                String output = sessionService.generateReportForSession(currentSession);
+                view.showData(output);
+                view.notify(reportPrintService.printReport(output, view.getCurrentClientIPAddress()));
+            } else {
+                log.warn("{} tried to close till but didn't have an open session", view.getCurrentUsername());
+                view.notify("No till session open");
+            }
+
         }
     }